do less #ifdef-ing
